refactor: extract SEXP recovery into general-purpose r_memory module
Move SEXPREC_DATA_OFFSET computation and try_recover_r_sexp() out of
arrow_impl.rs into a new r_memory module. Always initialized at
package init (not gated behind arrow feature).

R uses VECTOR_SEXPREC (48 bytes on 64-bit) for ALL vector types
(REALSXP, INTSXP, RAWSXP, STRSXP, VECSXP). Data starts at
sizeof(SEXPREC_ALIGN) after the SEXP address. Offset measured at
runtime to be portable across R versions and platforms.

Available for Arrow SEXP recovery, allocator validation, and future
Cow<[T]> round-trip optimization.
